+++ client/CharacterUtils.py
@@ -1,0 +1,8 @@
+class CharacterUtils:
+
+    carriage_return = '\r'
+    line_feed = '\n'
+    space = ' '
+    horizontal_tab = '\t'
+    crlf = carriage_return + line_feed
+    separator_characters = '()<>@,;:\\\"/[]?={}' + space + horizontal_tab
